autogen-tdd/requirements.txt

# AutoGen 멀티 에이전트 시스템 - 의존성 패키지

# ============================================================================
# 필수 패키지 (꼭 설치해야 함)
# ============================================================================

# AutoGen 핵심 패키지
autogen-agentchat>=0.4.0
autogen-ext>=0.4.0
autogen-ext[openai]  # OpenAI 모델도 사용하고 싶다면

# 환경 설정
python-dotenv>=1.0.0

# TDD 바이브코딩 시스템용 (pytest 필수)
pytest>=7.0.0
pytest-cov>=4.0.0

# 데이터 처리
asyncio-throttle>=1.0.2
pydantic>=2.0.0

# ============================================================================
# 개발 도구 (선택사항 - 코드 품질 향상용)
# ============================================================================

# 코드 포맷팅 및 린팅
black>=22.0.0
isort>=5.0.0
flake8>=6.0.0

# 타입 체크
mypy>=1.0.0

# 추가 pytest 플러그인
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# ============================================================================
# 유틸리티 (선택사항 - 사용성 개선용)
# ============================================================================

# 터미널 UI 개선
# rich>=13.0.0
# tqdm>=4.64.0

# 설정 파일 지원
# pyyaml>=6.0.0
# toml>=0.10.2

# 웹 인터페이스 (향후 확장용)
# fastapi>=0.100.0
# uvicorn>=0.22.0

# ============================================================================
# 빠른 설치 가이드
# ============================================================================

# 🚀 기본 설치 (바로 사용 가능):
# pip install autogen-agentchat autogen-ext python-dotenv pytest pytest-cov

# 📦 전체 설치 (개발 도구 포함):
# pip install -r requirements.txt

# 🔧 개별 설치:
# pip install autogen-agentchat autogen-ext  # AutoGen 핵심
# pip install python-dotenv                  # 환경 변수
# pip install pytest pytest-cov              # TDD 시스템용

# ============================================================================
# 시스템 요구사항
# ============================================================================
# - Python 3.8 이상
# - Gemini API 키 필요 (https://aistudio.google.com/app/apikey)
# - 가상환경 사용 권장